# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/studies": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Create Study",
        "description": "Create a new study in the active workspace.",
        "operationId": "create_study_api_admin_studies_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "List Studies",
        "description": "List studies in the active workspace.",
        "operationId": "list_studies_api_admin_studies_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyRead"
                  },
                  "title": "Response List Studies Api Admin Studies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study",
        "description": "Get study details.",
        "operationId": "get_study_api_admin_studies__slug__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Update Study",
        "description": "Update study configuration.",
        "operationId": "update_study_api_admin_studies__slug__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Delete Study",
        "description": "Delete a study (Superuser only, and must be Archived).",
        "operationId": "delete_study_api_admin_studies__slug__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/validate": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Validate Study",
        "description": "Check if study is ready for activation.",
        "operationId": "validate_study_api_admin_studies__slug__validate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Validate Study Api Admin Studies  Slug  Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/state": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Change Study State",
        "description": "Change study state (Draft <-> Active <-> Closed <-> Archived).",
        "operationId": "change_study_state_api_admin_studies__slug__state_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          },
          {
            "name": "new_state",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StudyState"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/stats": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study Stats",
        "description": "Get aggregated study statistics.",
        "operationId": "get_study_stats_api_admin_studies__slug__stats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyStatsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StudyState": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "paused",
          "closed",
          "archived"
        ],
        "title": "StudyState",
        "description": "Enum for study lifecycle states."
      },
      "StudyStatsRead": {
        "properties": {
          "started_count": {
            "type": "integer",
            "title": "Started Count"
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count"
          },
          "completion_rate": {
            "type": "number",
            "title": "Completion Rate"
          },
          "median_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Duration Seconds"
          },
          "device_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Device Breakdown"
          }
        },
        "type": "object",
        "required": [
          "started_count",
          "completed_count",
          "completion_rate",
          "median_duration_seconds",
          "device_breakdown"
        ],
        "title": "StudyStatsRead",
        "description": "Schema for study statistics."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}